Avoid overwriting by-serial links with APFS volumes
APFS volumes receive their own /dev/disk entry, but share the serial
number with the underlying device, which causes links to be overwritten.
This is fixed by ignoring devices that have a media which is not in the
'IODeviceTree:/'. This heuristic seems to work for APFS, but might not
generalize.